자동차 경주 게임
추가 사항
1.showWinner 메소드 기능 분리
   - mostGoValue( 1등 자동차 전진값 저장), winnerCarList( 1등 자동차 이름 저장), showWinner(1등 자동차 출력)  기능으로 분리

수정 사항
1. Car 클래스의 getCarName(), separateCarNameState(String carName), divideCarName(Car car) 메소드 이전
   - RacingGame 클래스로 이전
   - inputCarName 메소드 public -> public static 수정

2. 1 번과정에 따라 RacingGame main 메소드 코드 수정
